- Bootstrapped configuration with the default DI provider and migrated the existing settings architecture to the new .net core settings/options architecture. - Db settings are now injected (the connection string location has changed) - Patched up issues with inconsistencies in IFileProvider implementation and got template scanning up and running.

+++ src/Cofoundry.Domain/Domain/PageModuleTypes/Commands/RegisterPageModuleTypesCommandHandler.cs
@@ -123,7 +123,11 @@ namespace Cofoundry.Domain
                                 .PageModuleTemplates
                                 .Where(mt => !fileDetails.Templates.Any(t => t.FileName.Equals(mt.FileName, StringComparison.OrdinalIgnoreCase)))
                                 .ToList();
-            _dbContext.PageModuleTypeTemplates.RemoveRange(templatesToDelete);
+
+            if (templatesToDelete.Any())
+            {
+                _dbContext.PageModuleTypeTemplates.RemoveRange(templatesToDelete);
+            }
 
             foreach (var fileTemplate in fileDetails.Templates)
             {
